Changing && in || line 214
previous line : 
if((init_row + row != 0) && (init_column + column != 0))
new line : 
if((init_row + row != 0) || (init_column + column != 0))

idea of the process : verifying if the selected cell is the same than the one for which we are searching clues 
problem with previous version : if the selected cell is in the same column or line than the one for which we are searching clues, they are not analysed. If it was intended, it could be an optimisation because the program still analyse the cells that were set aside using the row_analysis and the column_analysis, so it prevent analysing same cells with two functions ; it is for this reason that the program still worked. This optimisation was not intended, in case I re-use this function with another purpose it's better not to put the optimisation.